[CDEC-352] Remove Pos.Core.Class
