+++ core/src/Pos/Core.hs
@@ -1,4 +1,3 @@
-{-# OPTIONS_GHC -Wno-unused-imports   #-}
 {-# OPTIONS_GHC -Wno-dodgy-exports    #-}
 
 module Pos.Core
@@ -10,7 +9,6 @@ module Pos.Core
        , module Pos.Core.ProtocolConstants
        , module Pos.Core.Delegation
        , module Pos.Core.Constants
-       , module Pos.Core.Class
        , module Pos.Core.Update
        , module Pos.Core.Ssc
        , module Pos.Core.Genesis
@@ -18,29 +16,22 @@ module Pos.Core
        , module Pos.Core.Configuration
        , module Pos.Core.Context
        , module Pos.Core.Common
-       , module Pos.Core.Block.Blockchain
-       , module Pos.Core.Block.Genesis
-       , module Pos.Core.Block.Main
-       , module Pos.Core.Block.Union
+       , module Pos.Core.Block
        ) where
 
+import           Pos.Core.Block
+import           Pos.Core.Common
+import           Pos.Core.Configuration
+import           Pos.Core.Constants
+import           Pos.Core.Context
+import           Pos.Core.Delegation
+import           Pos.Core.Genesis
+import           Pos.Core.ProtocolConstants
+import           Pos.Core.Script ()
+import           Pos.Core.Slotting
+import           Pos.Core.Ssc
+import           Pos.Core.Txp
+import           Pos.Core.Update
 import           Pos.Exception
 import           Pos.Merkle
 import           Pos.System.Metrics.Constants
-import           Pos.Core.Txp
-import           Pos.Core.Script ()
-import           Pos.Core.ProtocolConstants
-import           Pos.Core.Delegation
-import           Pos.Core.Constants
-import           Pos.Core.Class
-import           Pos.Core.Update
-import           Pos.Core.Ssc
-import           Pos.Core.Genesis
-import           Pos.Core.Slotting
-import           Pos.Core.Configuration
-import           Pos.Core.Context
-import           Pos.Core.Common
-import           Pos.Core.Block.Blockchain
-import           Pos.Core.Block.Genesis
-import           Pos.Core.Block.Main
-import           Pos.Core.Block.Union
