+++ core/src/Pos/Core/Block/Union/Instances.hs
@@ -19,12 +19,12 @@ import           Pos.Binary.Class (Bi)
 import           Pos.Core.Block.Blockchain (GenericBlock (..))
 import           Pos.Core.Block.Genesis ()
 import           Pos.Core.Block.Main ()
-import           Pos.Core.Block.Union.Types (Block, BlockHeader (..), blockHeaderHash,
+import           Pos.Core.Block.Union.Types (Block, BlockHeader (..), HasHeaderHash (..),
+                                             HasPrevBlock (..), IsGenesisHeader, IsHeader,
+                                             IsMainHeader (..), blockHeaderHash,
                                              choosingBlockHeader)
-import           Pos.Core.Class (HasDifficulty (..), HasEpochIndex (..), HasEpochOrSlot (..),
-                                 HasHeaderHash (..), HasPrevBlock (..), IsGenesisHeader, IsHeader,
-                                 IsMainHeader (..))
-import           Pos.Core.Slotting (EpochOrSlot (..))
+import           Pos.Core.Common (HasDifficulty (..))
+import           Pos.Core.Slotting (EpochOrSlot (..), HasEpochIndex (..), HasEpochOrSlot (..))
 import           Pos.Util.Some (Some)
 
 ----------------------------------------------------------------------------
