+++ core/src/Pos/Core/Block/Union/Types.hs
@@ -35,6 +35,11 @@ module Pos.Core.Block.Union.Types
        , blockHeaderHash
        , blockHeaderProtocolMagic
 
+       -- * IsHeader classes
+       , IsHeader
+       , IsGenesisHeader
+       , IsMainHeader (..)
+
        , module Pos.Core.Block.Genesis.Types
        , module Pos.Core.Block.Main.Types
        ) where
@@ -53,12 +58,12 @@ import           Pos.Core.Block.Blockchain (Blockchain (..), GenericBlock (..),
                                             GenericBlockHeader (..), gbHeader, gbhPrevBlock)
 import           Pos.Core.Block.Genesis.Types
 import           Pos.Core.Block.Main.Types
-import           Pos.Core.Common (ChainDifficulty)
+import           Pos.Core.Common (ChainDifficulty, HasDifficulty (..))
 import           Pos.Core.Delegation (ProxySigHeavy, ProxySigLight)
-import           Pos.Core.Slotting (SlotId (..))
+import           Pos.Core.Slotting (HasEpochIndex (..), HasEpochOrSlot (..), SlotId (..))
 import           Pos.Core.Ssc (mkSscProof)
 import           Pos.Core.Txp (mkTxProof)
-import           Pos.Core.Update (mkUpdateProof)
+import           Pos.Core.Update (HasBlockVersion (..), HasSoftwareVersion (..), mkUpdateProof)
 import           Pos.Crypto (Hash, ProtocolMagic, PublicKey, Signature, hash, unsafeHash)
 import           Pos.Util.Some (Some, applySome, liftLensSome)
 
@@ -264,3 +269,112 @@ blockHeaderProtocolMagic (BlockHeaderGenesis gbh) = _gbhProtocolMagic gbh
 blockHeaderProtocolMagic (BlockHeaderMain mbh)    = _gbhProtocolMagic mbh
 
 makePrisms 'BlockHeaderGenesis
+
+----------------------------------------------------------------------------
+-- IsHeader
+----------------------------------------------------------------------------
+
+{- | A class that lets subpackages use some fields from headers without
+depending on cardano-sl:
+
+  * 'difficultyL'
+  * 'epochIndexL'
+  * 'epochOrSlotG'
+  * 'prevBlockL'
+  * 'headerHashG'
+-}
+class ( HasDifficulty header
+      , HasEpochIndex header
+      , HasEpochOrSlot header
+      , HasPrevBlock header
+      , HasHeaderHash header) =>
+      IsHeader header
+
+instance HasDifficulty (Some IsHeader) where
+    difficultyL = liftLensSome difficultyL
+
+instance HasEpochIndex (Some IsHeader) where
+    epochIndexL = liftLensSome epochIndexL
+
+instance HasEpochOrSlot (Some IsHeader) where
+    getEpochOrSlot = applySome getEpochOrSlot
+
+instance HasPrevBlock (Some IsHeader) where
+    prevBlockL = liftLensSome prevBlockL
+
+instance HasHeaderHash (Some IsHeader) where
+    headerHash = applySome headerHash
+
+instance IsHeader (Some IsHeader)
+
+----------------------------------------------------------------------------
+-- IsGenesisHeader
+----------------------------------------------------------------------------
+
+-- | A class for genesis headers.
+class IsHeader header => IsGenesisHeader header
+
+instance HasDifficulty (Some IsGenesisHeader) where
+    difficultyL = liftLensSome difficultyL
+
+instance HasEpochIndex (Some IsGenesisHeader) where
+    epochIndexL = liftLensSome epochIndexL
+
+instance HasEpochOrSlot (Some IsGenesisHeader) where
+    getEpochOrSlot = applySome getEpochOrSlot
+
+instance HasPrevBlock (Some IsGenesisHeader) where
+    prevBlockL = liftLensSome prevBlockL
+
+instance HasHeaderHash (Some IsGenesisHeader) where
+    headerHash = applySome headerHash
+
+instance IsHeader        (Some IsGenesisHeader)
+instance IsGenesisHeader (Some IsGenesisHeader)
+
+----------------------------------------------------------------------------
+-- IsMainHeader
+----------------------------------------------------------------------------
+
+{- | A class for main headers. In addition to 'IsHeader', provides:
+
+  * 'headerSlotL'
+  * 'headerLeaderKeyL'
+  * 'blockVersionL'
+  * 'softwareVersionL'
+-}
+class (IsHeader header
+      ,HasBlockVersion header
+      ,HasSoftwareVersion header) =>
+      IsMainHeader header
+  where
+    -- | Id of the slot for which this block was generated.
+    headerSlotL :: Lens' header SlotId
+    -- | Public key of slot leader.
+    headerLeaderKeyL :: Lens' header PublicKey
+
+instance HasDifficulty (Some IsMainHeader) where
+    difficultyL = liftLensSome difficultyL
+
+instance HasEpochIndex (Some IsMainHeader) where
+    epochIndexL = liftLensSome epochIndexL
+
+instance HasEpochOrSlot (Some IsMainHeader) where
+    getEpochOrSlot = applySome getEpochOrSlot
+
+instance HasPrevBlock (Some IsMainHeader) where
+    prevBlockL = liftLensSome prevBlockL
+
+instance HasHeaderHash (Some IsMainHeader) where
+    headerHash = applySome headerHash
+
+instance HasBlockVersion (Some IsMainHeader) where
+    blockVersionL = liftLensSome blockVersionL
+
+instance HasSoftwareVersion (Some IsMainHeader) where
+    softwareVersionL = liftLensSome softwareVersionL
+
+instance IsHeader     (Some IsMainHeader)
+instance IsMainHeader (Some IsMainHeader) where
+    headerSlotL = liftLensSome headerSlotL
+    headerLeaderKeyL = liftLensSome headerLeaderKeyL
