--- core/src/Pos/Core/Class.hs
@@ -1,132 +1,0 @@
-{-# LANGUAGE CPP #-}
-{-# OPTIONS_GHC -Wno-redundant-constraints #-}
-
-module Pos.Core.Class
-       (
-       -- * Classes for overloaded accessors
-         HasPrevBlock (..)
-       , HasDifficulty (..)
-       , isMoreDifficult
-       , HasBlockVersion (..)
-       , HasSoftwareVersion (..)
-       , HasHeaderHash (..)
-       , HasEpochIndex (..)
-       , HasEpochOrSlot (..)
-       , epochOrSlotG
-
-       -- * Classes for headers
-       , IsHeader
-       , IsGenesisHeader
-       , IsMainHeader (..)
-       ) where
-
-import           Universum
-
-import           Pos.Core.Block.Union.Types (HasHeaderHash (..), HasPrevBlock (..))
-import           Pos.Core.Common (ChainDifficulty)
-import           Pos.Core.Slotting (HasEpochIndex (..), HasEpochOrSlot (..), SlotId, epochOrSlotG)
-import           Pos.Core.Update (BlockVersion, SoftwareVersion)
-import           Pos.Crypto.Signing (PublicKey)
-import           Pos.Util.Some (Some, applySome, liftLensSome)
-
-#define SOME_LENS_CLASS(HAS, LENS, CL)                       \
-    instance HAS (Some CL) where LENS = liftLensSome LENS
-#define SOME_FUNC_CLASS(HAS, FUNC, CL)                       \
-    instance HAS (Some CL) where FUNC = applySome FUNC
-
-----------------------------------------------------------------------------
--- Classes for overloaded accessors
-----------------------------------------------------------------------------
-
--- HasDifficulty
-class HasDifficulty a where
-    difficultyL :: Lens' a ChainDifficulty
-
-SOME_LENS_CLASS(HasDifficulty, difficultyL, HasDifficulty)
-
-isMoreDifficult :: HasDifficulty a => a -> a -> Bool
-a `isMoreDifficult` b = a ^. difficultyL > b ^. difficultyL
-
--- HasBlockVersion
-class HasBlockVersion a where
-    blockVersionL :: Lens' a BlockVersion
-
-SOME_LENS_CLASS(HasBlockVersion, blockVersionL, HasBlockVersion)
-
--- HasSoftwareVersion
-class HasSoftwareVersion a where
-    softwareVersionL :: Lens' a SoftwareVersion
-
-SOME_LENS_CLASS(HasSoftwareVersion, softwareVersionL, HasSoftwareVersion)
-
-----------------------------------------------------------------------------
--- Classes for headers
-----------------------------------------------------------------------------
-
--- Add (..) to export list when IsHeader or IsGenesisHeader get any methods
-
-{- | A class that lets subpackages use some fields from headers without
-depending on cardano-sl:
-
-  * 'difficultyL'
-  * 'epochIndexL'
-  * 'epochOrSlotG'
-  * 'prevBlockL'
-  * 'headerHashG'
--}
-class ( HasDifficulty header
-      , HasEpochIndex header
-      , HasEpochOrSlot header
-      , HasPrevBlock header
-      , HasHeaderHash header) =>
-      IsHeader header
-
-SOME_LENS_CLASS(HasDifficulty,  difficultyL,    IsHeader)
-SOME_LENS_CLASS(HasEpochIndex,  epochIndexL,    IsHeader)
-SOME_FUNC_CLASS(HasEpochOrSlot, getEpochOrSlot, IsHeader)
-SOME_LENS_CLASS(HasPrevBlock,   prevBlockL,     IsHeader)
-SOME_FUNC_CLASS(HasHeaderHash,  headerHash,     IsHeader)
-
-instance IsHeader (Some IsHeader)
-
--- | A class for genesis headers.
-class IsHeader header => IsGenesisHeader header
-
-SOME_LENS_CLASS(HasDifficulty,  difficultyL,    IsGenesisHeader)
-SOME_LENS_CLASS(HasEpochIndex,  epochIndexL,    IsGenesisHeader)
-SOME_FUNC_CLASS(HasEpochOrSlot, getEpochOrSlot, IsGenesisHeader)
-SOME_LENS_CLASS(HasPrevBlock,   prevBlockL,     IsGenesisHeader)
-SOME_FUNC_CLASS(HasHeaderHash,  headerHash,     IsGenesisHeader)
-
-instance IsHeader        (Some IsGenesisHeader)
-instance IsGenesisHeader (Some IsGenesisHeader)
-
-{- | A class for main headers. In addition to 'IsHeader', provides:
-
-  * 'headerSlotL'
-  * 'headerLeaderKeyL'
-  * 'blockVersionL'
-  * 'softwareVersionL'
--}
-class (IsHeader header
-      ,HasBlockVersion header
-      ,HasSoftwareVersion header) =>
-      IsMainHeader header
-  where
-    -- | Id of the slot for which this block was generated.
-    headerSlotL :: Lens' header SlotId
-    -- | Public key of slot leader.
-    headerLeaderKeyL :: Lens' header PublicKey
-
-SOME_LENS_CLASS(HasDifficulty,      difficultyL,      IsMainHeader)
-SOME_LENS_CLASS(HasEpochIndex,      epochIndexL,      IsMainHeader)
-SOME_FUNC_CLASS(HasEpochOrSlot,     getEpochOrSlot,   IsMainHeader)
-SOME_LENS_CLASS(HasPrevBlock,       prevBlockL,       IsMainHeader)
-SOME_FUNC_CLASS(HasHeaderHash,      headerHash,       IsMainHeader)
-SOME_LENS_CLASS(HasBlockVersion,    blockVersionL,    IsMainHeader)
-SOME_LENS_CLASS(HasSoftwareVersion, softwareVersionL, IsMainHeader)
-
-instance IsHeader     (Some IsMainHeader)
-instance IsMainHeader (Some IsMainHeader) where
-    headerSlotL = liftLensSome headerSlotL
-    headerLeaderKeyL = liftLensSome headerLeaderKeyL
